Simplify How Spells are Registered
Create Spell in SpellInit, specify its name, SpellType, and its action
Create SpellBook in SpellBookInit, specify its name, and Spell

Actions are stored in SpellActions
