+++ src/main/java/com/divinity/hlspells/HLSpells.java
@@ -1,9 +1,6 @@
 package com.divinity.hlspells;
 
-import com.divinity.hlspells.events.spell.HeldSpells;
-import com.divinity.hlspells.misc.CastSpells;
 import com.divinity.hlspells.setup.RegistryHandler;
-import com.divinity.hlspells.setup.client.ClientSetup;
 import com.divinity.hlspells.villages.POIFixup;
 import com.divinity.hlspells.villages.StructureGen;
 import com.divinity.hlspells.villages.Villagers;
@@ -38,8 +35,6 @@ public class HLSpells
 
         // Register ourselves for server and other game events we are interested in
         MinecraftForge.EVENT_BUS.addListener(this::setupMageHouses);
-        MinecraftForge.EVENT_BUS.register(new HeldSpells());
-        MinecraftForge.EVENT_BUS.register(new CastSpells());
         MinecraftForge.EVENT_BUS.register(this);
     }
