+++ src/main/java/com/divinity/hlspells/spell/Spell.java
@@ -5,25 +5,31 @@ import com.divinity.hlspells.init.SpellInit;
 import com.google.common.collect.Maps;
 import net.minecraft.entity.ai.attributes.Attribute;
 import net.minecraft.entity.ai.attributes.AttributeModifier;
+import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.Util;
 import net.minecraft.util.text.ITextComponent;
 import net.minecraft.util.text.TranslationTextComponent;
+import net.minecraft.world.World;
 import net.minecraftforge.registries.ForgeRegistryEntry;
 
 import javax.annotation.Nullable;
 import java.util.Map;
+import java.util.function.BiConsumer;
+import java.util.function.Consumer;
 
 public class Spell extends ForgeRegistryEntry<Spell> {
 
     private final Map<Attribute, AttributeModifier> attributeModifiers = Maps.newHashMap();
     private final SpellType category;
+    private final BiConsumer<PlayerEntity, World> spellAction;
 
     @Nullable
     private String descriptionId;
 
-    public Spell(SpellType spellType) {
+    public Spell(SpellType spellType, BiConsumer<PlayerEntity, World> spellAction) {
         this.category = spellType;
+        this.spellAction = spellAction;
     }
 
     public boolean isInstantenous() {
@@ -61,5 +67,9 @@ public class Spell extends ForgeRegistryEntry<Spell> {
 
     public static String getId(Spell spell) {
         return SpellInit.SPELLS_REGISTRY.get().toString();
+    }
+
+    public BiConsumer<PlayerEntity, World> getSpellAction() {
+        return spellAction;
     }
 }
