+++ src/main/java/com/divinity/hlspells/spell/SpellBookObject.java
@@ -2,12 +2,15 @@ package com.divinity.hlspells.spell;
 
 import com.divinity.hlspells.init.SpellBookInit;
 import com.google.common.collect.ImmutableList;
+import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.tags.ITag;
 import net.minecraft.util.ResourceLocation;
+import net.minecraft.world.World;
 import net.minecraftforge.registries.ForgeRegistryEntry;
 
 import javax.annotation.Nullable;
 import java.util.List;
+import java.util.function.Predicate;
 
 public class SpellBookObject extends ForgeRegistryEntry<SpellBookObject> {
 
@@ -52,5 +55,17 @@ public class SpellBookObject extends ForgeRegistryEntry<SpellBookObject> {
             }
         }
         return false;
+    }
+
+    public boolean containsSpell(Predicate<SpellInstance> predicate) {
+        return spellInstances.stream().anyMatch(predicate);
+    }
+
+    public boolean isEmpty() {
+        return this == SpellBookInit.EMPTY.get();
+    }
+
+    public void runAction(PlayerEntity entity, World world) {
+        spellInstances.forEach(spellInstance -> spellInstance.getSpell().getSpellAction().accept(entity, world));
     }
 }
